+++ CHANGELOG.md
@@ -27,6 +27,7 @@ stream
 - [#327](https://github.com/influxdata/kapacitor/issues/327): You can now window based on count in addition to time.
 - [#913](https://github.com/influxdata/kapacitor/issues/913): Add fillPeriod option to Window node, so that the first emit waits till the period has elapsed before emitting.
 - [#898](https://github.com/influxdata/kapacitor/issues/898): Now when the Window node every value is zero, the window will be emitted immediately for each new point.
+- [#1052](https://github.com/influxdata/kapacitor/issues/1052): Move alerta api token to header and add option to skip TLS verification.
 
 ### Bugfixes